_hurd_internal_post_signal: Split into more functions
* hurd/hurdsig.c (_hurd_internal_post_signal): Use inner functions
instead of gotos for the overall flow control.

This patch should change nothing besides replacing gotos with function
calls and return statements.

Note that the "signo == 0" test at the beginning of post_signal, is now
done on every call.  However, only the first call ever has a zero signo,
so this does not change the overall behavior.